[PimcoreBundle] move serializedData editable to PimcoreBundle
