+++ src/CoreShop/Bundle/PimcoreBundle/DependencyInjection/Configuration.php
@@ -45,6 +45,8 @@ final class Configuration implements ConfigurationInterface
                         ->ignoreExtraKeys(false)
                         ->children()
                             ->scalarNode('broker')->defaultValue('/bundles/coreshoppimcore/pimcore/js/broker.js')->end()
+                            ->scalarNode('core_extension_tag_serializedData')->defaultValue('/bundles/coreshoppimcore/pimcore/js/coreExtension/tags/coreShopSerializedData.js')->end()
+                            ->scalarNode('core_extension_data_serializedData')->defaultValue('/bundles/coreshoppimcore/pimcore/js/coreExtension/data/coreShopSerializedData.js')->end()
                         ->end()
                     ->end()
                     ->arrayNode('css')
