+++ src/CoreShop/Bundle/PimcoreBundle/Resources/public/pimcore/css/pimcore.css
@@ -1,0 +1,4 @@
+.coreshop_icon_serialized,
+.pimcore_icon_coreShopSerializedData {
+    background: url(../img/icons/serialized.svg) center center no-repeat !important;
+}
