+++ src/CoreShop/Bundle/PimcoreBundle/Resources/translations/admin.de.yml
@@ -1,0 +1,1 @@
+coreshop_serialized_data: 'Serialisierte Daten'
