+++ src/CoreShop/Bundle/PimcoreBundle/Resources/translations/admin.en.yml
@@ -1,0 +1,1 @@
+coreshop_serialized_data: 'Serialized Data'
