+++ src/CoreShop/Bundle/ResourceBundle/CoreExtension/SerializedData.php
@@ -12,188 +12,14 @@
 
 namespace CoreShop\Bundle\ResourceBundle\CoreExtension;
 
-use Pimcore\Model;
-
-class SerializedData extends Model\DataObject\ClassDefinition\Data
-{
+if (class_exists(\CoreShop\Bundle\PimcoreBundle\CoreExtension\SerializedData::class)) {
+    @trigger_error('Class CoreShop\Bundle\ResourceBundle\CoreExtension\SerializedData is deprecated since version 2.0.0-beta.2 and will be removed in 2.0. Use CoreShop\Bundle\PimcoreBundle\CoreExtension\SerializedData class instead.', E_USER_DEPRECATED);
+} else {
     /**
-     * Static type of this element
-     *
-     * @var string
+     * @deprecated Class CoreShop\Bundle\ResourceBundle\CoreExtension\SerializedData is deprecated since version 2.0.0-beta.2 and will be removed in 2.0. Use CoreShop\Bundle\PimcoreBundle\CoreExtension\SerializedData class instead.
      */
-    public $fieldtype = 'SerializedData';
-
-    /**
-     * Type for the column to query
-     *
-     * @var array
-     */
-    public $queryColumnType = null;
-
-    /**
-     * Type for the column
-     *
-     * @var array
-     */
-    public $columnType = 'LONGBLOB';
-
-    /**
-     * Type for the generated phpdoc
-     *
-     * @var string
-     */
-    public $phpdocType;
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getDataForResource($data, $object = null, $params = [])
-    {
-        return serialize($data);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getDataFromResource($data, $object = null, $params = [])
-    {
-        return unserialize($data) ?: null;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getDataForQueryResource($data, $object = null, $params = [])
-    {
-        return null;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getDataForEditmode($data, $object = null, $params = [])
-    {
-        return $data;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getDataFromEditmode($data, $object = null, $params = [])
-    {
-        return $this->getDataFromResource($data, $object, $params);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getDataFromGridEditor($data, $object = null, $params = [])
-    {
-        return $data;
-    }
-
-    /**
-     * @return string
-     */
-    public function getQueryColumnType()
-    {
-        return null;
-    }
-
-    /**
-     * @return string
-     */
-    public function getColumnType()
-    {
-        return 'LONGBLOB';
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function checkValidity($data, $omitMandatoryCheck = false)
-    {
-        return true;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function isEmpty($data)
-    {
-        return is_null($data);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getForWebserviceExport($object, $params = [])
-    {
-        return null;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getFromWebserviceImport($value, $object = null, $params = [], $idMapper = null)
-    {
-        // not implemented
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getDataForGrid($data, $object = null, $params = [])
-    {
-        return $this->getDataFromResource($data, $object, $params);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getVersionPreview($data, $object = null, $params = [])
-    {
-        return $this->getDataFromResource($data, $object, $params);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getForCsvExport($object, $params = [])
-    {
-        return null;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getFromCsvImport($importValue, $object = null, $params = [])
+    class SerializedData
     {
 
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getFilterCondition($value, $operator, $params = [])
-    {
-        return null;
-    }
-
-    /**
-     * @param array $data
-     *
-     * @return $this
-     */
-    public function setValues($data = [])
-    {
-        foreach ($data as $key => $value) {
-            $method = 'set'.$key;
-            if (method_exists($this, $method)) {
-                $this->$method($value);
-            }
-        }
-
-        return $this;
     }
 }
