+++ src/CoreShop/Bundle/ResourceBundle/Resources/config/pimcore/config.yml
@@ -5,14 +5,6 @@ jms_serializer:
                 namespace_prefix: "Pimcore"
                 path: "@CoreShopResourceBundle/Resources/config/serializer"
 
-pimcore:
-    objects:
-        class_definitions:
-            data:
-                map:
-                    coreShopSerializedData: CoreShop\Bundle\ResourceBundle\CoreExtension\SerializedData
-
-
 doctrine:
     dbal:
         types:
